pyproject.toml
[tool.poetry]
name = "django-quotes"
version = "0.1.0"
description = """A reusable Django app to collect quotes for use in random retrieval or generation of sentences using \
    Markov Chains."""
authors = ["Daniel Andrlik <daniel@andrlik.org>"]
license = "BSD-3-Clause"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
django-model-utils = "^4.2.0"
rules = "^3.1"
Markdown = "^3.3.6"
markovify = "^0.9.3"
spacy = "^3.2.2"
python-slugify = "^6.1.1"
Sphinx = {version = "^4.4.0", optional=true}
sphinx-autobuild = {version = "^2021.3.14", optional=true}
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
drf-spectacular = "^0.21.2"
django-crispy-forms = "^1.14.0"
loguru = "^0.6.0"

[tool.poetry.dependencies.en_core_web_sm]
url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz"

[tool.poetry.dev-dependencies]
Werkzeug = {extras = ["watchdog"], version = "^2.0.2"}
ipdb = "^0.13.9"
psycopg2-binary = "^2.9.3"
watchgod = "^0.7"
mypy = "^0.931"
django-stubs = "^1.9.0"
pytest = "^7.0.0"
pytest-sugar = "^0.9.4"
djangorestframework-stubs = "^1.4.0"
Sphinx = "^4.4.0"
sphinx-autobuild = "^2021.3.14"
flake8 = "^4.0.1"
flake8-isort = "^4.1.1"
coverage = "^6.3.1"
black = {extras = ["d"], version = "^22.1.0"}
pylint-django = "^2.5.0"
pre-commit = "^2.17.0"
factory-boy = "^3.2.1"
django-debug-toolbar = "^3.2.4"
django-extensions = "^3.1.5"
django-coverage-plugin = "^2.0.2"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
sphinx-rtd-theme = "^1.0.0"
django-easy-logging = "^0.4.0"

[tool.poetry.extras]
docs = ["sphinx", "sphinx-autobuild"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
